+++ safari/src/main/java/com/gd/safari/vo/ProfileImg.java
@@ -9,7 +9,7 @@ public class ProfileImg {
 	String fileExt;
 	String originName;
 	String fileType;
-	int fileSize;
+	long fileSize;
 	String createDate;
 	String updateDate;
